+++ platform/android/src/native_map_view.cpp
@@ -61,12 +61,9 @@ NativeMapView::NativeMapView(jni::JNIEnv& _env,
                              const jni::Object<NativeMapView>& _obj,
                              const jni::Object<FileSource>& jFileSource,
                              const jni::Object<MapRenderer>& jMapRenderer,
-                             jni::jfloat _pixelRatio,
-                             jni::jboolean _crossSourceCollisions)
-    : javaPeer(_env, _obj)
-    , mapRenderer(MapRenderer::getNativePeer(_env, jMapRenderer))
-    , pixelRatio(_pixelRatio) {
-
+                             jni::jfloat pixelRatio_,
+                             jni::jboolean crossSourceCollisions_)
+    : javaPeer(_env, _obj), mapRenderer(MapRenderer::getNativePeer(_env, jMapRenderer)), pixelRatio(pixelRatio_) {
     // Get a reference to the JavaVM for callbacks
     if (_env.GetJavaVM(&vm) < 0) {
         _env.ExceptionDescribe();
@@ -79,11 +76,11 @@ NativeMapView::NativeMapView(jni::JNIEnv& _env,
     // Create Map options
     MapOptions options;
     options.withMapMode(MapMode::Continuous)
-           .withSize(mbgl::Size{ static_cast<uint32_t>(width), static_cast<uint32_t>(height) })
-           .withPixelRatio(pixelRatio)
-           .withConstrainMode(ConstrainMode::HeightOnly)
-           .withViewportMode(ViewportMode::Default)
-           .withCrossSourceCollisions(_crossSourceCollisions);
+        .withSize(mbgl::Size{static_cast<uint32_t>(width), static_cast<uint32_t>(height)})
+        .withPixelRatio(pixelRatio)
+        .withConstrainMode(ConstrainMode::HeightOnly)
+        .withViewportMode(ViewportMode::Default)
+        .withCrossSourceCollisions(crossSourceCollisions_);
 
     // Create the core map
     map = std::make_unique<mbgl::Map>(
@@ -595,10 +592,6 @@ void NativeMapView::setDebug(JNIEnv&, jni::jboolean debug) {
     map->setDebug(debugOptions);
 }
 
-void NativeMapView::cycleDebugOptions(JNIEnv&) {
-    map->cycleDebugOptions();
-}
-
 jni::jboolean NativeMapView::getDebug(JNIEnv&) {
     return map->getDebug() != DebugOptions::NoDebug;
 }
@@ -625,8 +618,60 @@ jni::Local<jni::Object<PointF>> NativeMapView::pixelForLatLng(JNIEnv& env, jdoub
     return PointF::New(env, static_cast<float>(pixel.x), static_cast<float>(pixel.y));
 }
 
+void NativeMapView::pixelsForLatLngs(JNIEnv& env,
+                                     const jni::Array<jdouble>& input,
+                                     jni::Array<jdouble>& output,
+                                     jfloat pixelRatio_) {
+    jni::NullCheck(env, &input);
+    std::size_t len = input.Length(env);
+
+    std::vector<mbgl::LatLng> latLngs;
+    latLngs.reserve(len);
+
+    for (std::size_t i = 0; i < len; i += 2) {
+        auto latLng = mbgl::LatLng(input.Get(env, i), input.Get(env, i + 1));
+        latLngs.push_back(latLng);
+    }
+
+    std::vector<jdouble> buffer;
+    buffer.reserve(len);
+    std::vector<ScreenCoordinate> coordinates = map->pixelsForLatLngs(latLngs);
+    for (std::size_t i = 0; i < len / 2; i++) {
+        buffer.push_back(coordinates[i].x * pixelRatio_);
+        buffer.push_back(coordinates[i].y * pixelRatio_);
+    }
+
+    output.SetRegion<std::vector<jdouble>>(env, 0, buffer);
+}
+
 jni::Local<jni::Object<LatLng>> NativeMapView::latLngForPixel(JNIEnv& env, jfloat x, jfloat y) {
     return LatLng::New(env, map->latLngForPixel(mbgl::ScreenCoordinate(x, y)));
+}
+
+void NativeMapView::latLngsForPixels(JNIEnv& env,
+                                     const jni::Array<jdouble>& input,
+                                     jni::Array<jdouble>& output,
+                                     jfloat pixelRatio_) {
+    jni::NullCheck(env, &input);
+    std::size_t len = input.Length(env);
+
+    std::vector<mbgl::ScreenCoordinate> coordinates;
+    coordinates.reserve(len);
+
+    for (std::size_t i = 0; i < len; i += 2) {
+        auto coordinate = mbgl::ScreenCoordinate(input.Get(env, i) / pixelRatio_, input.Get(env, i + 1) / pixelRatio_);
+        coordinates.push_back(coordinate);
+    }
+
+    std::vector<jdouble> buffer;
+    buffer.reserve(len);
+    std::vector<mbgl::LatLng> latLngs = map->latLngsForPixels(coordinates);
+    for (std::size_t i = 0; i < len / 2; i++) {
+        buffer.push_back(latLngs[i].latitude());
+        buffer.push_back(latLngs[i].longitude());
+    }
+
+    output.SetRegion<std::vector<jdouble>>(env, 0, buffer);
 }
 
 jni::Local<jni::Array<jlong>> NativeMapView::addPolylines(JNIEnv& env, const jni::Array<jni::Object<Polyline>>& polylines) {
@@ -1120,15 +1165,16 @@ void NativeMapView::registerNative(jni::JNIEnv& env) {
         METHOD(&NativeMapView::updateMarker, "nativeUpdateMarker"),
         METHOD(&NativeMapView::addMarkers, "nativeAddMarkers"),
         METHOD(&NativeMapView::setDebug, "nativeSetDebug"),
-        METHOD(&NativeMapView::cycleDebugOptions, "nativeCycleDebugOptions"),
         METHOD(&NativeMapView::getDebug, "nativeGetDebug"),
         METHOD(&NativeMapView::isFullyLoaded, "nativeIsFullyLoaded"),
         METHOD(&NativeMapView::onLowMemory, "nativeOnLowMemory"),
         METHOD(&NativeMapView::getMetersPerPixelAtLatitude, "nativeGetMetersPerPixelAtLatitude"),
         METHOD(&NativeMapView::projectedMetersForLatLng, "nativeProjectedMetersForLatLng"),
         METHOD(&NativeMapView::pixelForLatLng, "nativePixelForLatLng"),
+        METHOD(&NativeMapView::pixelsForLatLngs, "nativePixelsForLatLngs"),
         METHOD(&NativeMapView::latLngForProjectedMeters, "nativeLatLngForProjectedMeters"),
         METHOD(&NativeMapView::latLngForPixel, "nativeLatLngForPixel"),
+        METHOD(&NativeMapView::latLngsForPixels, "nativeLatLngsForPixels"),
         METHOD(&NativeMapView::addPolylines, "nativeAddPolylines"),
         METHOD(&NativeMapView::addPolygons, "nativeAddPolygons"),
         METHOD(&NativeMapView::updatePolyline, "nativeUpdatePolyline"),