+++ render-test/manifest_parser.cpp
@@ -34,8 +34,18 @@ const std::vector<std::pair<std::string, std::string>>& Manifest::getIgnores() c
 const std::string& Manifest::getTestRootPath() const {
     return testRootPath;
 }
+const std::string& Manifest::getAssetPath() const {
+    return assetPath;
+}
 const std::string& Manifest::getManifestPath() const {
     return manifestPath;
+}
+const std::string& Manifest::getResultPath() const {
+    return resultPath;
+}
+
+const std::set<std::string>& Manifest::getProbes() const {
+    return probes;
 }
 
 void Manifest::doShuffle(uint32_t seed) {
@@ -217,18 +227,16 @@ std::vector<std::pair<std::string, std::string>> parseIgnores(const std::vector<
     return ignores;
 }
 
-std::vector<mbgl::filesystem::path> getTestExpectations(mbgl::filesystem::path testPath,
-                                                        const mbgl::filesystem::path& testsRootPath,
-                                                        std::vector<mbgl::filesystem::path> expectationsPaths) {
-    std::vector<mbgl::filesystem::path> expectations{std::move(testPath.remove_filename())};
-    const auto& defaultTestExpectationsPath = expectations.front().string();
-
-    const std::regex regex{testsRootPath.string()};
-    for (const auto& path : expectationsPaths) {
-        expectations.emplace_back(std::regex_replace(defaultTestExpectationsPath, regex, path.string()));
-        assert(!expectations.back().empty());
+// defaultExpectationPath: absolute path that constains the style.json file for testing
+// testId: Test case id that used for composing expectation path
+// expectatedPaths: absolute paths that constain possible expected.png/metrics.json files for result checking
+std::vector<mbgl::filesystem::path> getTestExpectations(const mbgl::filesystem::path& defaultExpectationPath,
+                                                        const std::string& testId,
+                                                        std::vector<mbgl::filesystem::path> expectatedPaths) {
+    std::vector<mbgl::filesystem::path> expectations{defaultExpectationPath};
+    for (const auto& expectedPath : expectatedPaths) {
+        expectations.emplace_back(expectedPath / testId);
     }
-
     return expectations;
 }
 
@@ -239,7 +247,7 @@ mbgl::filesystem::path getValidPath(const std::string& manifestPath, const std::
         result = BasePath / result;
     }
     if (mbgl::filesystem::exists(result)) {
-        return result;
+        return result.lexically_normal();
     }
     mbgl::Log::Warning(mbgl::Event::General, "Invalid path is provoided inside the manifest file: %s", path.c_str());
     return mbgl::filesystem::path{};
@@ -249,7 +257,7 @@ mbgl::filesystem::path getValidPath(const std::string& manifestPath, const std::
 
 mbgl::optional<Manifest> ManifestParser::parseManifest(const std::string& manifestPath,
                                                        const std::vector<std::string>& testNames,
-                                                       const std::string& testFilter) {
+                                                       std::string testFilter) {
     Manifest manifest;
     const auto filePath = mbgl::filesystem::path(manifestPath);
     manifest.manifestPath = manifestPath.substr(0, manifestPath.find(filePath.filename()));
@@ -285,6 +293,18 @@ mbgl::optional<Manifest> ManifestParser::parseManifest(const std::string& manife
             return mbgl::nullopt;
         }
     }
+    if (document.HasMember("result_path")) {
+        const auto& resultPathValue = document["result_path"];
+        if (!resultPathValue.IsString()) {
+            mbgl::Log::Warning(
+                mbgl::Event::General, "Invalid assetPath is provoided inside the manifest file: %s", filePath.c_str());
+            return mbgl::nullopt;
+        }
+        manifest.resultPath = (getValidPath(manifest.manifestPath, resultPathValue.GetString()) / "").string();
+        if (manifest.resultPath.empty()) {
+            return mbgl::nullopt;
+        }
+    }
     mbgl::filesystem::path baseTestPath;
     if (document.HasMember("base_test_path")) {
         const auto& testPathValue = document["base_test_path"];
@@ -312,6 +332,20 @@ mbgl::optional<Manifest> ManifestParser::parseManifest(const std::string& manife
             return mbgl::nullopt;
         }
         enbaleProbeTest = true;
+    }
+    mbgl::filesystem::path expectedMetricPath;
+    if (document.HasMember("metric_path")) {
+        const auto& metricPathValue = document["metric_path"];
+        if (!metricPathValue.IsString()) {
+            mbgl::Log::Warning(mbgl::Event::General,
+                               "Invalid metric_path is provoided inside the manifest file: %s",
+                               filePath.c_str());
+            return mbgl::nullopt;
+        }
+        expectedMetricPath = getValidPath(manifest.manifestPath, metricPathValue.GetString());
+        if (expectedMetricPath.empty()) {
+            return mbgl::nullopt;
+        }
     }
     std::vector<mbgl::filesystem::path> expectationPaths{};
     if (document.HasMember("expectation_paths")) {
@@ -359,12 +393,47 @@ mbgl::optional<Manifest> ManifestParser::parseManifest(const std::string& manife
         manifest.ignores = parseIgnores(ignorePaths);
     }
 
+    if (document.HasMember("probes")) {
+        const auto& probesValue = document["probes"];
+        if (!probesValue.IsArray()) {
+            mbgl::Log::Warning(mbgl::Event::General,
+                               "Provided probes inside the manifest file: %s is not a valid array",
+                               filePath.c_str());
+            return mbgl::nullopt;
+        }
+        for (const auto& value : probesValue.GetArray()) {
+            if (!value.IsString()) {
+                mbgl::Log::Warning(mbgl::Event::General,
+                                   "Invalid probe type is provoided inside the manifest file: %s",
+                                   filePath.c_str());
+                return mbgl::nullopt;
+            }
+            manifest.probes.emplace(value.GetString());
+        }
+    }
+
+    if (testFilter.empty() && document.HasMember("filter")) {
+        const auto& filterValue = document["filter"];
+        if (!filterValue.IsString()) {
+            mbgl::Log::Warning(
+                mbgl::Event::General, "Invalid filter is provoided inside the manifest file: %s", filePath.c_str());
+            return mbgl::nullopt;
+        }
+
+        testFilter = filterValue.GetString();
+    }
+
     manifest.testRootPath = enbaleProbeTest ? probeTestPath.string() : baseTestPath.string();
     if (manifest.testRootPath.back() == '/') {
         manifest.testRootPath.pop_back();
     }
     if (manifest.manifestPath.back() == '/') {
         manifest.manifestPath.pop_back();
+    }
+    if (manifest.resultPath.empty()) {
+        manifest.resultPath = manifest.manifestPath;
+    } else if (manifest.resultPath.back() == '/') {
+        manifest.resultPath.pop_back();
     }
 
     std::vector<mbgl::filesystem::path> paths;
@@ -391,7 +460,18 @@ mbgl::optional<Manifest> ManifestParser::parseManifest(const std::string& manife
             }
 
             if (testPath.path().filename() == "style.json") {
-                testPaths.emplace_back(testPath, getTestExpectations(testPath, path, expectationPaths));
+                const auto defaultExpectationPath{std::move(mbgl::filesystem::path(testPath).remove_filename())};
+                const auto rootLength = manifest.testRootPath.length();
+                auto testId = defaultExpectationPath.string();
+                testId = testId.substr(rootLength + 1, testId.length() - rootLength - 1);
+
+                std::vector<mbgl::filesystem::path> expectedMetricPaths{expectedMetricPath};
+#if defined(__ANDROID__)
+                expectedMetricPaths.emplace_back("/sdcard/");
+#endif
+                testPaths.emplace_back(testPath,
+                                       getTestExpectations(defaultExpectationPath, testId, expectationPaths),
+                                       getTestExpectations(defaultExpectationPath, testId, expectedMetricPaths));
             }
         }
     }