+++ render-test/manifest_parser.hpp
@@ -17,7 +17,10 @@ public:
     const std::vector<std::pair<std::string, std::string>>& getIgnores() const;
     const std::vector<TestPaths>& getTestPaths() const;
     const std::string& getTestRootPath() const;
+    const std::string& getAssetPath() const;
     const std::string& getManifestPath() const;
+    const std::string& getResultPath() const;
+    const std::set<std::string>& getProbes() const;
     void doShuffle(uint32_t seed);
 
     std::string localizeURL(const std::string& url) const;
@@ -43,13 +46,15 @@ private:
     std::string testRootPath;
     std::string vendorPath;
     std::string assetPath;
+    std::string resultPath;
     std::vector<std::pair<std::string, std::string>> ignores;
     std::vector<TestPaths> testPaths;
+    std::set<std::string> probes;
 };
 
 class ManifestParser {
 public:
     static mbgl::optional<Manifest> parseManifest(const std::string& manifestPath,
                                                   const std::vector<std::string>& testNames,
-                                                  const std::string& testFilter);
+                                                  std::string testFilter);
 };